[MOS-686] Fixed MTP availability only after phone unlocked 2
This is fix for Windows. It's fix for previous fix wich:
Fixed file access via MTP even when phone is not unlocked.
Now access is granted when the phone is unlocked by the user entering
a passcode. If the phone is not passcode protected (passcode is nor set)
then access to the files is always possible via MTP.