[CP-2184] Fix issue with reset events getting lost
* Fix of the issue that reset events in some cases
would get lost if previous event wasn't dequeued
before the next one came. Currently if the
queue is already full when trying to push
new event, the previous one is considered
outdated and thus invalid and gets overwritten.
* Fix sending uninitialized events value in
case usbDeviceStateCB was called for event
that OS doesn't need to handle.